+++ kernel/src/arch/x86/PIT.c
@@ -1,7 +1,7 @@
 #include <arch/x86/PIT.h>
 #include <lib/ISR_IRQ.h>
 #include <arch/x86/defs/IRQ.h>
-#include <lib/io.h>
+#include <arch/x86/io.h>
 #include <lib/std.h>
 #include <bios/vga.h>
